linux-uconsole-cm3-rpi64: expose DSI status though sysfs
Due to a potential bug(raspberrypi/linux issue 4323) in VC4's driver, the DSI
panel cannot be reliably initialized.

The updated patch set fix the poweroff sequence of the DSI display, but
also expose the bug to post-power-up environment.

When testing the DSI bus by turning the panel on and off, there's a 10
percent chance that error occurs when initializing the panel.

I believe Clockworkpi has recognized this issue before the first batch
of uConsole coming out, as there are big trunks of code disabled by
preprocessor directives. Their solution is to only initialize the
display once at boot.

The panel driver is updated to properly deinitialize the panel, but that
makes the bug having a chance to occur when the screen is
unblanked(recover from idle black screen).

In order to workaround it, the DSI bus error state is exposed through a
sysfs node. On RPi3 series, it's `/sys/devices/platform/soc/3f700000.dsi/3f700000.dsi.0/dsi_state`.
That is the `dsi_state` file in the panel node's folder. Read it and `ok\n`
means no error and `error\n` means error occurred. Userspace programs can
read this file and blank/unblank the screen when needed.

There are different methods to blank/unblank the display under different
circumstances. There's one known method to restart the display under any
environment: write 1 then 0 to `/sys/class/graphics/fb0/blank`. `fb0` is
the framebuffer node for the DSI panel.

A service script is to be written as an example userspace program to
workaround the issue.